Minor revision to the "from_range" function. One key optimization for the case where step=1 (uses run container). Otherwise, makes API standard in the sense that all intervals are [min,max) which is the default in all our code. Also minor simplification and cleaning memory leaks.

+++ include/containers/array.h
@@ -73,9 +73,9 @@ static inline bool array_container_nonzero_cardinality(
 /* Copy one container into another. We assume that they are distinct. */
 void array_container_copy(const array_container_t *src, array_container_t *dst);
 
-/*  Add all the values between min and max (included) at a distance k*step from min.
+/*  Add all the values in [min,max) (included) at a distance k*step from min.
     The container must have a size less to 4096 after this addition. */
-void array_container_add_from_range(array_container_t *arr, uint16_t min, uint16_t max,
+void array_container_add_from_range(array_container_t *arr, uint32_t min, uint32_t max,
                                     uint16_t step);
 
 /* Set the cardinality to zero (does not release memory). */
